Fix handling of errno invoking external functions
If an external function in KLEE is invoked it might update errno.
This patch explicitly updates errno of the external process state before
invoking the external function with the execution state's value.
After invocation, the execution state's errno value is updated again.

The new implementation tries to reduce the time between creation and
use of the value by explicitly preserving the errno value.

* Add support for `errno()` for Darwin as well.
* Simplified errno handling in POSIX layer